Congdvph28728

+++ bookingapi/src/main/java/com/poly/bookingapi/entity/Discount.java
@@ -25,7 +25,7 @@ public class Discount {
     private String nameDiscount;
 
     @Column(name = "discount_value")
-    private Double discountValue;
+    private Integer discountValue;
 
     @Column(name = "status")
     private Integer status;
